+++ example/src/scenes/Ipc/index.js
@@ -15,6 +15,8 @@ import IpcProgressBar from './IpcProgressBar';
 import IpcTimerInterval from './IpcTimerInterval';
 import IpcPtz from './IpcPtz';
 import IpcDragSort from './IpcDragSort';
+import IpcMessagePlayer from './IpcMessagePlayer';
+import IpcGpsSignal from './IpcGpsSignal';
 
 export default class AnimatedScene extends Component {
   get data() {
@@ -34,6 +36,8 @@ export default class AnimatedScene extends Component {
   static IpcTimerInterval = IpcTimerInterval;
   static IpcPtz = IpcPtz;
   static IpcDragSort = IpcDragSort;
+  static IpcMessagePlayer = IpcMessagePlayer;
+  static IpcGpsSignal = IpcGpsSignal;
 
   render() {
     return <TYFlatList contentContainerStyle={{ paddingTop: 16 }} data={this.data} />;